add update product deploy api
