+++ pkg/microservice/aslan/core/environment/service/revision.go
@@ -20,12 +20,13 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/koderover/zadig/pkg/microservice/aslan/core/common/service/render"
+
 	"go.uber.org/zap"
 
 	commonmodels "github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/models"
 	commonrepo "github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/mongodb"
 	templaterepo "github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/mongodb/template"
-	commonservice "github.com/koderover/zadig/pkg/microservice/aslan/core/common/service"
 	"github.com/koderover/zadig/pkg/microservice/aslan/core/common/service/kube"
 	commonutil "github.com/koderover/zadig/pkg/microservice/aslan/core/common/util"
 	"github.com/koderover/zadig/pkg/setting"
@@ -133,12 +134,12 @@ func GetProductRevision(product *commonmodels.Product, allServiceTmpls []*common
 	var newRender, oldRender *commonmodels.RenderSet
 	if prodTmpl.ProductFeature == nil || prodTmpl.ProductFeature.DeployType == setting.K8SDeployType {
 		// TODO is it the right way to fetch new product?
-		newRender, err = commonservice.GetRenderSet(product.Render.Name, 0, false, product.EnvName, log)
+		newRender, err = render.GetRenderSet(product.Render.Name, 0, false, product.EnvName, log)
 		if err != nil {
 			return prodRev, err
 		}
 
-		oldRender, err = commonservice.GetRenderSet(product.Render.Name, product.Render.Revision, false, product.EnvName, log)
+		oldRender, err = render.GetRenderSet(product.Render.Name, product.Render.Revision, false, product.EnvName, log)
 		if err != nil {
 			return prodRev, err
 		}
